requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2
python-dateutil==2.8.2
pytz==2023.3
yfinance==0.2.28
polygon-api-client==1.13.0
pandas==2.1.4
numpy==1.26.2
openai>=1.0.0

